+++ src/dashboard.ts
@@ -2,18 +2,13 @@ import { readFileSync, existsSync, unlinkSync, rmSync } from 'node:fs';
 import { resolve } from 'node:path';
 import { execFile, ChildProcessPromise } from 'promisify-child-process';
 
-import { TimedCacheValidity, Database, DatabaseConfig, Source } from './db.js';
+import { Database, DatabaseConfig, Source } from './db.js';
 import { startRequest, completeRequest } from './perf-tracker.js';
 import { BenchmarkCompletion } from './api.js';
 import { GitHub } from './github.js';
 import { robustPath, siteConfig } from './util.js';
 import { assert, log } from './logging.js';
-import {
-  CompareView,
-  WarmupData,
-  WarmupDataForTrial,
-  WarmupDataPerCriterion
-} from './views/view-types.js';
+import { CompareView } from './views/view-types.js';
 import { prepareCompareView } from './stats-data-prep.js';
 
 const reportOutputFolder = resolve(robustPath(`../resources/reports/`));
@@ -24,193 +19,6 @@ JOIN Experiment exp ON exp.sourceId = s.id
 JOIN Measurement m ON  m.expId = exp.id
 WHERE repoURL = 'https://github.com/smarr/ReBenchDB'
  */
-
-export async function dashProfile(
-  runId: number,
-  trialId: number,
-  db: Database
-): Promise<any> {
-  const result = await db.query({
-    name: 'fetchProfileDataByRunIdTrialId',
-    text: `
-          SELECT substring(commitId, 1, 6) as commitid,
-            benchmark.name as bench, executor.name as exe, suite.name as suite,
-            cmdline, varValue, cores, inputSize, extraArgs,
-            invocation, numIterations, warmup, value as profile
-          FROM ProfileData
-            JOIN Trial ON trialId = Trial.id
-            JOIN Experiment ON expId = Experiment.id
-            JOIN Source ON source.id = sourceId
-            JOIN Run ON runId = run.id
-            JOIN Suite ON suiteId = suite.id
-            JOIN Benchmark ON benchmarkId = benchmark.id
-            JOIN Executor ON execId = executor.id
-          WHERE runId = $1 AND trialId = $2`,
-    values: [runId, trialId]
-  });
-
-  const data = result.rows[0];
-  try {
-    data.profile = JSON.parse(data.profile);
-  } catch (e) {
-    /* let's just leave it as a string */
-  }
-  return data;
-}
-
-export async function dashMeasurements(
-  projectSlug: string,
-  runId: number,
-  trialId1: number,
-  trialId2: number,
-  db: Database
-): Promise<WarmupData | null> {
-  const q = {
-    name: 'fetchMeasurementsByProjectIdRunIdTrialId',
-    text: `SELECT
-              trialId,
-              invocation, iteration, warmup,
-              criterion.name as criterion,
-              criterion.unit as unit,
-              value
-            FROM
-              Measurement
-              JOIN Trial ON trialId = Trial.id
-              JOIN Experiment ON Trial.expId = Experiment.id
-              JOIN Criterion ON criterion = criterion.id
-              JOIN Run ON runId = run.id
-              JOIN Project ON Project.id = Experiment.projectId
-            WHERE Project.slug = $1
-              AND runId = $2
-              AND (trialId = $3 OR trialId = $4)
-            ORDER BY trialId, criterion, invocation, iteration;`,
-    values: [projectSlug, runId, trialId1, trialId2]
-  };
-  const result = await db.query(q);
-  if (result.rows.length === 0) {
-    return null;
-  }
-
-  const trial1: WarmupDataForTrial = {
-    trialId: result.rows[0].trialid,
-    warmup: result.rows[0].warmup,
-    data: []
-  };
-  const trial2: WarmupDataForTrial = {
-    trialId: result.rows[result.rows.length - 1].trialid,
-    warmup: result.rows[result.rows.length - 1].warmup,
-    data: []
-  };
-  // preprocess rows, but should already have it like this in the database...
-
-  let trialId = 0;
-  let currentTrial: WarmupDataForTrial | null = null;
-  let lastCriterion = null;
-  let critObject: WarmupDataPerCriterion | null = null;
-
-  for (const r of result.rows) {
-    if (trialId === 0) {
-      trialId = r.trialid;
-      currentTrial = trial1;
-      lastCriterion = null;
-    } else if (trialId !== r.trialid) {
-      if (currentTrial === trial2) {
-        throw Error(
-          'Unexpected trialId change. We only expect two different ones.'
-        );
-      }
-      trialId = r.trialid;
-      currentTrial = trial2;
-      lastCriterion = null;
-    }
-
-    if (lastCriterion === null || lastCriterion !== r.criterion) {
-      lastCriterion = r.criterion;
-      critObject = {
-        criterion: r.criterion,
-        unit: r.unit,
-        values: []
-      };
-      currentTrial?.data.push(critObject);
-    }
-
-    if (critObject) {
-      if (critObject.values[r.invocation - 1] === undefined) {
-        critObject.values[r.invocation - 1] = [];
-      }
-
-      critObject.values[r.invocation - 1][r.iteration - 1] = r.value;
-    }
-  }
-
-  return { trial1, trial2 };
-}
-
-let statisticsCache: { stats: any[] } | null = null;
-let statsCacheValid: TimedCacheValidity | null = null;
-
-export function statsCache(): TimedCacheValidity | null {
-  return statsCacheValid;
-}
-
-export async function dashStatistics(db: Database): Promise<{ stats: any[] }> {
-  if (
-    statisticsCache !== null &&
-    statsCacheValid !== null &&
-    statsCacheValid.isValid()
-  ) {
-    return statisticsCache;
-  }
-
-  statsCacheValid = db.getStatsCacheValidity();
-
-  const result = await db.query({
-    name: 'fetchStats',
-    text: `
-        SELECT * FROM (
-          SELECT 'Experiments' as table, count(*) as cnt FROM experiment
-          UNION ALL
-          SELECT 'Trials' as table, count(*) as cnt FROM trial
-          UNION ALL
-          SELECT 'Executors' as table, count(*) as cnt FROM executor
-          UNION ALL
-          SELECT 'Benchmarks' as table, count(*) as cnt FROM benchmark
-          UNION ALL
-          SELECT 'Projects' as table, count(*) as cnt FROM project
-          UNION ALL
-          SELECT 'Suites' as table, count(*) as cnt FROM suite
-          UNION ALL
-          SELECT 'Environments' as table, count(*) as cnt FROM environment
-          UNION ALL
-          SELECT 'Runs' as table, count(*) as cnt FROM run
-          UNION ALL
-          SELECT 'Measurements' as table, count(*) as cnt FROM measurement
-        ) as counts
-        ORDER BY counts.table`
-  });
-  statisticsCache = { stats: result.rows };
-  return statisticsCache;
-}
-
-export async function dashChanges(
-  projectId: number,
-  db: Database
-): Promise<{ changes: any[] }> {
-  const result = await db.query({
-    name: 'fetchAllChangesByProjectId',
-    text: ` SELECT commitId, branchOrTag, projectId, repoURL, commitMessage,
-                max(startTime) as experimentTime
-            FROM experiment
-            JOIN Trial ON expId = experiment.id
-            JOIN Source ON sourceId = source.id
-            JOIN Project ON projectId = project.id
-            WHERE project.id = $1
-            GROUP BY commitId, branchOrTag, projectId, repoURL, commitMessage
-            ORDER BY max(startTime) DESC`,
-    values: [projectId]
-  });
-  return { changes: result.rows };
-}
 
 export async function dashDataOverview(
   projectId: number,
